feat: Car 파일 분리
