+++ src/Car.js
@@ -1,0 +1,10 @@
+export class Car { 
+  constructor(name) {
+      this.name = name;
+      this.location = 0;
+  }
+
+  move() {
+      this.location += 1;
+  }
+}
