+++ src/main.js
@@ -1,10 +1,1 @@
-export class Car { 
-    constructor(name) {
-        this.name = name;
-        this.location = 0;
-    }
-
-    move() {
-        this.location += 1;
-    }
-}
+console.log('main')
